+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [7.0.0] - 2024-06-27
+
 This release contains **BREAKING CHANGES**. See the [v7.0 Upgrade Guide][breaking7] for required
 changes you may need to make.
 
@@ -961,7 +963,8 @@ See the Python Migration Guide for details on how to update your YAML configurat
 - Nearly all command line options removed in favor of YAML equivalents.
 - Completely removed old python project & source code
 
-[Unreleased]: https://github.com/recyclarr/recyclarr/compare/v6.0.2...HEAD
+[Unreleased]: https://github.com/recyclarr/recyclarr/compare/v7.0.0...HEAD
+[7.0.0]: https://github.com/recyclarr/recyclarr/compare/v6.0.2...v7.0.0
 [6.0.2]: https://github.com/recyclarr/recyclarr/compare/v6.0.1...v6.0.2
 [6.0.1]: https://github.com/recyclarr/recyclarr/compare/v6.0.0...v6.0.1
 [6.0.0]: https://github.com/recyclarr/recyclarr/compare/v5.4.3...v6.0.0